Update search functionality and remove ExportGenerateCommand
This commit primarily enhances the search functionality by adding product category sorting for variants. Furthermore, it includes improvements such as adding locale to filters and normalizing locale placeholders. The ExportGenerateCommand file has been removed, as it was deemed unnecessary. The model Result has undergone a minor revision in the hits field. The version of dependencies has been updated as well.