Add SuggestProductTransformer Remove unused method in ProductMappingService
A SuggestProductTransformer has been added that attaches a ProductEntity to the SuggestItem based on the product number. This entity is then used to display the product price in the suggest result. Additionally, an unused method in the ProductMappingService has been removed, since it had been previously moved to the ProductUtil.